Move OpenMP functions out of Kokkos_OpenMP_Instance.hpp
